Implemented Dragon Smash Quick Build for Crux Prime Dragons
Implemented a script for the crux prime dragons to spawn the Golem Quick Build that insta kills the Entity if the player hits the dragon after the golem is built.  Tested on Local Ubuntu instance against 1 Crux Prime Maelstrom Dragon and 1 Butterscorch and had no issues building the quick build or with the enemy not smashing.  Enemies also correctly reset to their default behavior after the stun period has expired.