Add support to name and describe models
Adds support to name and describe models.
Tested that User Generated Content models can be named and described and that the property does not get its name overwritten anymore.
Regular models maintain their name and description across instances.
This does not allow the name to persist when the model is picked up as this is a larger implementation.
Tested that taming pets and spawning pets does not cause crashes or issues.