Added spawn management commands, along with some minor bugfixes.
Added commands to print to chat, despawn, and save to a bug reports entry, lists of entities, based on entity groups, or who/what spawned them.

Added a command to despawn an object by object id, and one to despawn based on the 'nearest to player' check done by the inspect command.

Added a command to print a player's object id to chat.

Also replaced (splitArgs.empty()) checks with (args == ""), because the argument parser never returns an empty vector, only a vector containing an empty string.

Updated Commands.md to reflect the added commands.